reconfigured the y plates, shrunk the y axis offset.
The motor shafts aren't long enough to span two MFD sheets, so
make a cutout in one, mount motor in the other.

Also started working on the main belt tensioner, in the
extrusion_mounts.scad - plan to use eye bolts, springs, a sliding
rod mount and a beefy seat at the end.

Paul