Full prototype on the belt tensioner ready :-)
Using eye bolts tied to springs on the rod sliders, with a brass
knurled nut tensioning.  A small oversight, I need to make a bed
mount!  Right now the idea is just to use a simple lasercut panel,
so I've put off the design cuz it's easy and the lasercutter is
far away.